pyproject.toml
[project]
name = "finances-reporting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "binance-connector>=3.12.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
